Update ext.publisher_low_search.php
Need to check if a previously called extension has altered the value of `$params`. I know that Low uses `low_search_pre_search` to add support for Low Events to Low Search, and there may be other add-ons using this same hook so it needs to play nicely!

It's likely that the same is required for `low_search_update_index` but I have not tested.